requirements.txt
matplotlib==3.7.1
numpy==1.23.5
opencv_python==4.7.0.72
pandas==2.0.2
Pillow==9.5.0
Pillow==9.5.0
ruptures==1.0.3
seaborn==0.12.2
streamlit==1.22.0
streamlit_option_menu==0.3.6
tensorflow==2.12.0
tensorflow_intel==2.12.0
